Everything about flake-schemas must be in schema.rs
